+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Fixed
 -----
 - Make Resolwe work with Python 3 again
 - Fix tests
+- Render data name again after inputs are resolved
 
 
 ==================